Move classes not related to store creation to different packages
