+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/onboarding/StoreOnboardingScreen.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.login.storecreation.onboarding
+package com.woocommerce.android.ui.onboarding
 
 import android.content.res.Configuration
 import androidx.compose.animation.core.animateFloatAsState
@@ -57,11 +57,11 @@ import com.woocommerce.android.R.string
 import com.woocommerce.android.ui.compose.component.Toolbar
 import com.woocommerce.android.ui.compose.component.WCTag
 import com.woocommerce.android.ui.compose.component.WCTextButton
-import com.woocommerce.android.ui.login.storecreation.onboarding.StoreOnboardingViewModel.AboutYourStoreTaskRes
-import com.woocommerce.android.ui.login.storecreation.onboarding.StoreOnboardingViewModel.Companion.NUMBER_ITEMS_IN_COLLAPSED_MODE
-import com.woocommerce.android.ui.login.storecreation.onboarding.StoreOnboardingViewModel.LaunchStoreTaskRes
-import com.woocommerce.android.ui.login.storecreation.onboarding.StoreOnboardingViewModel.OnboardingState
-import com.woocommerce.android.ui.login.storecreation.onboarding.StoreOnboardingViewModel.OnboardingTaskUi
+import com.woocommerce.android.ui.onboarding.StoreOnboardingViewModel.AboutYourStoreTaskRes
+import com.woocommerce.android.ui.onboarding.StoreOnboardingViewModel.Companion.NUMBER_ITEMS_IN_COLLAPSED_MODE
+import com.woocommerce.android.ui.onboarding.StoreOnboardingViewModel.LaunchStoreTaskRes
+import com.woocommerce.android.ui.onboarding.StoreOnboardingViewModel.OnboardingState
+import com.woocommerce.android.ui.onboarding.StoreOnboardingViewModel.OnboardingTaskUi
 
 @Composable
 fun StoreOnboardingScreen(viewModel: StoreOnboardingViewModel) {
