+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/onboarding/aboutyourstore/AboutYourStoreFragment.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.login.storecreation.onboarding.payments
+package com.woocommerce.android.ui.onboarding.aboutyourstore
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,11 +8,9 @@ import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
-import com.woocommerce.android.extensions.navigateSafely
 import com.woocommerce.android.ui.base.BaseFragment
 import com.woocommerce.android.ui.common.wpcomwebview.WPComWebViewAuthenticator
 import com.woocommerce.android.ui.compose.theme.WooThemeWithBackground
-import com.woocommerce.android.ui.login.storecreation.onboarding.payments.GetPaidViewModel.ShowWooPaymentsSetupSuccess
 import com.woocommerce.android.ui.main.AppBarStatus
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event.Exit
 import dagger.hilt.android.AndroidEntryPoint
@@ -20,14 +18,12 @@ import org.wordpress.android.fluxc.network.UserAgent
 import javax.inject.Inject
 
 @AndroidEntryPoint
-class GetPaidFragment : BaseFragment() {
-    private val viewModel: GetPaidViewModel by viewModels()
+class AboutYourStoreFragment : BaseFragment() {
+    private val viewModel: AboutYourStoreViewModel by viewModels()
 
-    @Inject
-    lateinit var userAgent: UserAgent
+    @Inject lateinit var userAgent: UserAgent
 
-    @Inject
-    lateinit var authenticator: WPComWebViewAuthenticator
+    @Inject lateinit var authenticator: WPComWebViewAuthenticator
 
     override val activityAppBarStatus: AppBarStatus
         get() = AppBarStatus.Hidden
@@ -37,7 +33,7 @@ class GetPaidFragment : BaseFragment() {
             setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
             setContent {
                 WooThemeWithBackground {
-                    GetPaidScreen(viewModel, userAgent, authenticator)
+                    AboutYourStoreScreen(viewModel, userAgent, authenticator)
                 }
             }
         }
@@ -51,10 +47,6 @@ class GetPaidFragment : BaseFragment() {
     private fun setupObservers() {
         viewModel.event.observe(viewLifecycleOwner) { event ->
             when (event) {
-                is ShowWooPaymentsSetupSuccess -> findNavController().navigateSafely(
-                    GetPaidFragmentDirections.actionGetPaidFragmentToWooPaymentsSetupCelebrationDialog()
-                )
-
                 is Exit -> findNavController().popBackStack()
             }
         }
