+++ WooCommerce/src/main/res/navigation/nav_graph_settings.xml
@@ -187,7 +187,7 @@
         android:label="CloseAccountDialogFragment" />
     <dialog
         android:id="@+id/nameYourStoreDialogFragment"
-        android:name="com.woocommerce.android.ui.login.storecreation.onboarding.NameYourStoreDialogFragment"
+        android:name="com.woocommerce.android.ui.onboarding.NameYourStoreDialogFragment"
         android:label="NameYourStoreDialogFragment">
         <argument
             android:name="fromOnboarding"
