+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/payments/taptopay/about/TapToPayAboutViewModelTest.kt
@@ -5,7 +5,7 @@ import com.woocommerce.android.cardreader.config.CardReaderConfigForCanada
 import com.woocommerce.android.cardreader.config.CardReaderConfigForGB
 import com.woocommerce.android.cardreader.config.CardReaderConfigForSupportedCountry
 import com.woocommerce.android.cardreader.config.CardReaderConfigForUSA
-import com.woocommerce.android.ui.login.storecreation.countrypicker.LocalCountriesRepository
+import com.woocommerce.android.ui.common.LocalCountriesRepository
 import com.woocommerce.android.util.CurrencyFormatter
 import com.woocommerce.android.viewmodel.BaseUnitTest
 import com.woocommerce.android.viewmodel.ResourceProvider
